[Clean] Split types and implementation codes into some files
