--- src/drop-line/index.ts
@@ -1,2 +1,0 @@
-export { getDropLineDirectionFromXY } from "./direciton";
-export { getDropLinePosition, checkIsInStackableArea } from "./position";
