+++ src/index.ts
@@ -1,2 +1,3 @@
-export * from "./list.component";
-export * from "./item.component";
+export { DragEndMeta, DragStartMeta, GhostRendererMeta, StackMeta, DropLineRendererInjectedProps } from "./list";
+export { List } from "./list.component";
+export { Item } from "./item.component";
