+++ src/list/meta/destinations.ts
@@ -1,0 +1,6 @@
+import { ItemIdentifier } from "../../shared";
+
+export type DestinationMeta<T extends ItemIdentifier> = {
+  groupIdentifier: T | undefined;
+  index: number | undefined;
+};
