+++ src/list/meta/stacks.ts
@@ -1,0 +1,5 @@
+import { ItemIdentifier, NodeMeta } from "../../shared";
+
+export type StackMeta<T extends ItemIdentifier> = Pick<NodeMeta<T>, "identifier" | "groupIdentifier" | "index" | "isGroup"> & {
+  nextGroupIdentifier: T | undefined;
+};
