--- src/node/index.ts
@@ -1,2 +1,0 @@
-export * from "./element";
-export * from "./node-meta";
