+++ src/shared/items/index.ts
@@ -1,0 +1,1 @@
+export { ItemIdentifier } from "./identifiers";
