+++ src/shared/nodes/index.ts
@@ -1,0 +1,2 @@
+export { ElementPosition } from "./elements";
+export { NodeMeta, getNodeMeta } from "./node-meta";
