+++ tsconfig.json
@@ -5,7 +5,6 @@
     "baseUrl": "./",
     "declaration": true,
     "esModuleInterop": true,
-    "isolatedModules": true,
     "jsx": "react",
     "lib": [
       "es2020",
